extract participanttable dialogues
