+++ client/src/pages/private/ParticipantTableDialogues.js
@@ -1,0 +1,160 @@
+import React, { useMemo } from 'react';
+import store from 'store';
+import {
+  ToastStatus,
+  InterviewingFormSchema,
+  RejectedFormSchema,
+  API_URL,
+  makeToasts,
+} from '../../constants';
+import { Dialog } from '../../components/generic';
+import {
+  ProspectingForm,
+  InterviewingForm,
+  RejectedForm,
+  HireForm,
+  NewParticipantForm,
+  EditParticipantForm,
+  ArchiveHiredParticipantForm,
+} from '../../components/modal-forms';
+import { useToast } from '../../hooks';
+import { getDialogTitle } from '../../utils';
+import { AuthContext, ParticipantsContext } from '../../providers';
+
+export const ParticipantTableDialogues = ({
+  forceReload,
+  setActiveModalForm,
+  activeModalForm,
+  actionMenuParticipant,
+  setActionMenuParticipant,
+}) => {
+  const { openToast } = useToast();
+  const { dispatch: participantsDispatch } = ParticipantsContext.useParticipantsContext();
+  const { auth } = AuthContext.useAuth();
+  const sites = useMemo(() => auth.user?.sites || [], [auth.user?.sites]);
+
+  const handleEngage = async (participantId, status, additional = {}) => {
+    const response = await fetch(`${API_URL}/api/v1/employer-actions`, {
+      method: 'POST',
+      headers: {
+        Authorization: `Bearer ${store.get('TOKEN')}`,
+        Accept: 'application/json',
+        'Content-type': 'application/json',
+      },
+      body: JSON.stringify({ participantId, status, data: additional }),
+    });
+
+    if (response.ok) {
+      const { data: statusData } = await response.json();
+
+      if (status === 'prospecting') {
+        // Modal appears after submitting
+        setActiveModalForm('prospecting');
+      } else {
+        const { firstName, lastName } = actionMenuParticipant;
+        const toasts = makeToasts(firstName, lastName);
+        openToast(toasts[statusData?.status === 'already_hired' ? statusData.status : status]);
+        setActionMenuParticipant(null);
+        setActiveModalForm(null);
+        forceReload();
+      }
+    } else {
+      openToast({
+        status: ToastStatus.Error,
+        message: response.error || response.statusText || 'Server error',
+      });
+    }
+  };
+
+  const onClose = () => {
+    setActiveModalForm(null);
+    setActionMenuParticipant(null);
+  };
+
+  return (
+    <Dialog
+      title={getDialogTitle(activeModalForm)}
+      open={activeModalForm != null}
+      onClose={onClose}
+    >
+      {activeModalForm === 'prospecting' && (
+        <ProspectingForm
+          name={`${actionMenuParticipant.firstName} ${actionMenuParticipant.lastName}`}
+          onClose={() => {
+            forceReload();
+            onClose();
+          }}
+          onSubmit={async () => {
+            onClose();
+
+            participantsDispatch({
+              type: ParticipantsContext.types.SELECT_TAB,
+              payload: 'My Candidates',
+            });
+          }}
+        />
+      )}
+
+      {activeModalForm === 'interviewing' && (
+        <InterviewingForm
+          initialValues={{ contactedDate: '' }}
+          validationSchema={InterviewingFormSchema}
+          onSubmit={(values) => {
+            handleEngage(actionMenuParticipant.id, 'interviewing', {
+              contacted_at: values.contactedDate,
+            });
+          }}
+          onClose={onClose}
+        />
+      )}
+
+      {activeModalForm === 'rejected' && (
+        <RejectedForm
+          initialValues={{ contactedDate: '' }}
+          validationSchema={RejectedFormSchema}
+          onSubmit={(values) => {
+            handleEngage(actionMenuParticipant.id, 'rejected', {
+              final_status: values.finalStatus,
+            });
+          }}
+          onClose={onClose}
+        />
+      )}
+
+      {activeModalForm === 'hired' && (
+        <HireForm
+          sites={sites}
+          onSubmit={(values) => {
+            handleEngage(actionMenuParticipant.id, 'hired', {
+              nonHcapOpportunity: values.nonHcapOpportunity,
+              positionTitle: values.positionTitle,
+              positionType: values.positionType,
+              hiredDate: values.hiredDate,
+              startDate: values.startDate,
+              site: values.site,
+            });
+          }}
+          onClose={onClose}
+        />
+      )}
+      {activeModalForm === 'edit-participant' && (
+        <EditParticipantForm
+          initialValues={actionMenuParticipant}
+          onClose={onClose}
+          submissionCallback={forceReload}
+        />
+      )}
+      {activeModalForm === 'new-participant' && (
+        <NewParticipantForm sites={sites} onClose={onClose} submissionCallback={forceReload} />
+      )}
+      {activeModalForm === 'archive' && (
+        <ArchiveHiredParticipantForm
+          onSubmit={(values) => {
+            handleEngage(actionMenuParticipant.id, 'archived', values);
+          }}
+          onClose={onClose}
+        />
+      )}
+    </Dialog>
+  );
+};
